Backend/src/requirements.txt
fastapi
uvicorn
minio
python-dotenv
scikit-learn
pandas
numpy
pytest
requests
pydantic
